scxtop: Fix panic in BPF event conversion
Replace .expect() with graceful error handling in edm.rs to prevent
crashes when encountering malformed BPF events. The tool now logs and
skips invalid events instead of panicking.

Reorganize test suite by moving EDM tests from inline cfg(test) modules
to a dedicated tests/ directory for better organization and
maintainability. Add new test case for error handling path to ensure
graceful degradation works correctly.

Fixes: Critical panic in production on unexpected BPF event types
